Extract email token login helpers into an import
